In: #13 (REFACTOR) Organize main page layout, streamline imports, and restructure FindByName - Simplified library imports using wildcard `*`. - Moved the `FindByName` file to UI package.
